refactor: reorganize & plan redesign
