+++ src/index.ts
@@ -1,5 +1,5 @@
 export { traceStorageAccess, Tracer } from "@/lib/trace";
-export { findLayoutInfoAtSlot } from "@/lib/slot-engine";
+export { findLayoutInfoAtSlot } from "@/lib/slots/engine";
 export type {
   StorageAccessTrace,
   TraceStorageAccessOptions,
