+++ src/lib/adapter/schema.ts
