+++ src/lib/slots/mapping.ts
@@ -1,9 +1,7 @@
-import { Address, Hex, hexToBigInt, isHex, keccak256, toHex } from "tevm";
-import { abi } from "@shazow/whatsabi";
-import { decodeAbiParameters, encodeAbiParameters, padHex } from "viem";
+import { Hex, keccak256 } from "tevm";
 
-import { isStorageAdapterType, StorageAdapter, StorageLayoutAdapter } from "@/lib/layout/adapter";
-import { AbiType, AbiTypeToPrimitiveType } from "@/lib/layout/schema";
+import { StorageAdapter, StorageLayoutAdapter } from "@/lib/adapter";
+import { AbiType } from "@/lib/adapter/schema";
 import { MappingKey, SlotLabelResult } from "@/lib/types";
 
 // Maximum nesting depth to prevent excessive computation
@@ -18,245 +16,12 @@ const NESTED_MAPPINGS_LIMIT = 4;
 const MAX_EXPLORED_STATES = 5_000;
 
 /**
- * A slot computation engine that implements Solidity's storage layout rules to accurately compute and label storage
- * slots.
- */
-
-/**
  * Computes the storage slot for a mapping given the base slot and key
  *
  * `keccak256(abi.encode(key, slot))
  */
 export const computeMappingSlot = (baseSlot: Hex, key: Hex): Hex =>
   keccak256(`0x${key.replace("0x", "")}${baseSlot.replace("0x", "")}`);
-
-/**
- * Computes the storage slot for a dynamic array element
- *
- * - `keccak256(slot) + index` for dynamic arrays
- * - `slot + index` for static arrays
- */
-export const computeArraySlot = (baseSlot: Hex, index: bigint | number | string, isDynamic = true): Hex => {
-  if (isDynamic) {
-    const baseSlotHash = keccak256(baseSlot);
-    const slotBigInt = hexToBigInt(baseSlotHash);
-    return toHex(slotBigInt + BigInt(index), { size: 32 });
-  } else {
-    const slotBigInt = hexToBigInt(baseSlot);
-    return toHex(slotBigInt + BigInt(index), { size: 32 });
-  }
-};
-
-/** Extract values from a transaction trace that might be used as keys or indices */
-export const extractPotentialKeys = (
-  trace: {
-    uniqueStackValues?: Array<string>;
-    relevantOps?: Array<{
-      op: string;
-      stack: Array<string>;
-    }>;
-  },
-  addresses: Array<Address>,
-  abiFunctions: Array<abi.ABIFunction>,
-  txData?: Hex,
-): MappingKey[] => {
-  const keys: MappingKey[] = [];
-
-  // Add touched addresses
-  addresses.forEach((address) => {
-    keys.push({
-      hex: padHex(address, { size: 32 }),
-      decoded: address,
-      type: "address",
-    });
-  });
-
-  // Extract parameters from transaction data
-  if (txData && txData.length > 10) {
-    const selector = txData.slice(0, 10);
-    const inputs = abiFunctions.find((fn) => fn.selector === selector)?.inputs;
-
-    if (inputs) {
-      // Decode function inputs
-      const params = decodeAbiParameters(inputs, `0x${txData.slice(10)}`);
-
-      params.forEach((param, index) => {
-        // If it's an array, add each element as a key
-        if (Array.isArray(param)) {
-          param.forEach((p) => {
-            const type = inputs[index].type.replace("[]", "") as AbiType;
-
-            if (type) {
-              keys.push({
-                hex: padHex(encodeAbiParameters([{ type }], [p]), { size: 32 }),
-                decoded: p as AbiTypeToPrimitiveType<typeof type>,
-                type,
-              });
-            }
-          });
-        } else {
-          // Otherwise just add the key straight up
-          const type = inputs[index].type as AbiType;
-
-          if (type) {
-            keys.push({
-              hex: padHex(encodeAbiParameters([{ type }], [param]), { size: 32 }),
-              decoded: param as AbiTypeToPrimitiveType<typeof type>,
-              type,
-            });
-          }
-        }
-      });
-    }
-  }
-
-  // Process stack values from the trace
-  if (trace.uniqueStackValues?.length) {
-    // Process unique stack values directly
-    for (const stackValue of trace.uniqueStackValues) {
-      keys.push({
-        hex: isHex(stackValue) ? padHex(stackValue, { size: 32 }) : toHex(stackValue, { size: 32 }),
-        type: undefined,
-      });
-    }
-  }
-
-  // Deduplicate keys
-  const uniqueMap = new Map();
-  // Add the new key only if it's not already in the map (and don't replace a key with a defined type)
-  keys.forEach((k) => {
-    if (!uniqueMap.has(k.hex) || k.type) uniqueMap.set(k.hex, k);
-  });
-
-  return Array.from(uniqueMap.values()).sort((a, b) => {
-    // prefer address as it's more likely to be a key
-    if (a.type === "address") return -1;
-    if (b.type === "address") return 1;
-    // prefer defined types
-    if (a.type === undefined) return 1;
-    if (b.type === undefined) return -1;
-
-    return 0;
-  });
-};
-
-/**
- * Finds all matching labels for a storage slot, including packed variables Using the StorageLayoutAdapter approach
- *
- * @param slot The storage slot to find information for
- * @param adapter A storage layout adapter
- * @param potentialKeys Potential mapping keys or array indices from the transaction
- * @returns Array of labeled slot results
- */
-export const findLayoutInfoAtSlot = (
-  slot: Hex,
-  adapter: StorageLayoutAdapter,
-  potentialKeys: MappingKey[],
-): SlotLabelResult[] => {
-  // Get all storage variables
-  const variables = Object.values(adapter);
-
-  // 1. Check for all direct variable matches at this slot (packed or unpacked)
-  // TODO: if it's a bytes type we're missing some data that is on the next slot, the whole thing is broken (maybe if that's the type iterate slots? the next slots will probably also be marked in the access list)
-  const directMatches = variables.filter((v) => v.getSlot() === slot);
-  if (directMatches.length > 0) {
-    // Return direct matches
-    return (
-      directMatches
-        // Sort by offset to ensure consistent order (helps when there are packed variables)
-        .sort((a, b) => (a.storageItem.offset ?? 0) - (b.storageItem.offset ?? 0))
-        .map((v) => ({
-          label: v.label,
-          slot,
-          matchType: "exact",
-          type: v.type,
-          offset: v.storageItem.offset,
-        }))
-    );
-  }
-
-  // 2. Check for mapping matches
-  const mappings = variables.filter(
-    isStorageAdapterType.mapping,
-  ) as StorageAdapter<`mapping(${string} => ${string})`>[];
-
-  for (const mapping of mappings) {
-    // Try to find a match using our unified function
-    const slotInfo = findMappingMatch(mapping.label, slot, mapping, potentialKeys, adapter);
-    if (slotInfo) return [slotInfo];
-  }
-
-  // TODO: 3. Check for array matches
-  // -> should actually provide the provider to adapters, so for arrays we can get the length (hope it was not manipulated), and get slots for all existing indexes (+1 in case an item was removed??)
-  // TODO: 4. Check for struct matches
-
-  // Process each variable in the adapter for arrays and other types
-  // for (const variableName of Object.keys(adapter)) {
-  //   try {
-  //     const variable = adapter[variableName];
-  //     if (!variable || !variable.type) continue;
-
-  //     // CASE 3: Array (static or dynamic)
-  //     if (variable.type.includes("[") && potentialKeys.length > 0) {
-  //       // For arrays, we need to check if the variable has itemSlot method
-  //       if ("getItemSlot" in variable) {
-  //         for (const key of potentialKeys) {
-  //           if (!key.hex) continue;
-
-  //           // Convert to numeric index if possible
-  //           let index: number;
-  //           if (key.type === "uint256" && typeof key.decoded === "bigint") {
-  //             index = Number(key.decoded);
-  //           } else if (isHex(key.hex)) {
-  //             index = Number(hexToBigInt(key.hex));
-  //           } else {
-  //             continue; // Not a valid index
-  //           }
-
-  //           // Skip invalid indices
-  //           if (isNaN(index) || index < 0 || index > 1_000_000) continue;
-
-  //           try {
-  //             // Get the slot for this array index
-  //             const itemSlot = (variable as any).getItemSlot({ index });
-
-  //             // Check if it matches our target
-  //             if (itemSlot === slot) {
-  //               // Extract the base type from the array type (e.g., uint256[] -> uint256)
-  //               const baseType = variable.type.match(/^([^\[]+)/)?.[1]?.trim() as AbiType;
-
-  //               results.push({
-  //                 label: `${variableName}[${index}]`,
-  //                 slot,
-  //                 matchType: "array",
-  //                 type: baseType || (variable.type as AbiType),
-  //                 index: key.hex,
-  //               });
-
-  //               break; // One match is enough
-  //             }
-  //           } catch (e) {
-  //             // Skip this index
-  //           }
-  //         }
-  //       }
-  //     }
-  //   } catch (e) {
-  //     // Skip this variable entirely
-  //     continue;
-  //   }
-  // }
-
-  // If no matches found, use generic variable name
-  return [
-    {
-      label: `var_${slot.slice(0, 10)}`,
-      slot: slot,
-      matchType: "exact",
-      type: undefined,
-    },
-  ];
-};
 
 /**
  * Finds a matching mapping key combination that produces the target storage slot. Works with both simple mappings and
